commits exception exercise

+++ src/util/InputTest.java
@@ -5,8 +5,8 @@ public class InputTest {
     public static void main(String[] args) {
         Input input = new Input();
 
-        input.getDouble(10,20);
-
+//        input.getDouble(10,20);
+        input.getInt(10,20);
 
     }
 }
